fix: resolve all clippy warnings
- Rename from_str methods to parse to avoid confusion with std::str::FromStr
- Add allow attributes for dead_code in API modules (tenant, relations, middleware)
- Fix unused imports across multiple modules
- Use sort_by_key instead of sort_by for cleaner code
- Fix doctest fn main removal
- Add thread-local-tenant feature to prax-query
- Rename Include::add to Include::with to avoid confusion with std::ops::Add
- Rename PooledBuffer::as_mut to as_mut_str to avoid AsMut confusion
- Add allow(clippy::should_implement_trait) for Filter::not
- Add allow(clippy::needless_range_loop) for performance-optimized loops
- Add allow(clippy::result_large_err) for prax-postgres error type
- Fix approximate PI constant in tests (use 3.5 instead of 3.14)
- Simplify URL masking in seed example to avoid url crate dependency
- Add allow(dead_code, unused) to examples and benchmarks

+++ prax-postgres/src/pool.rs
@@ -72,9 +72,9 @@ impl PgPool {
     pub fn status(&self) -> PoolStatus {
         let status = self.inner.status();
         PoolStatus {
-            available: status.available as usize,
-            size: status.size as usize,
-            max_size: status.max_size as usize,
+            available: status.available,
+            size: status.size,
+            max_size: status.max_size,
             waiting: status.waiting,
         }
     }
@@ -126,7 +126,7 @@ impl PgPool {
     pub async fn warmup(&self, count: usize) -> PgResult<()> {
         info!(count = count, "Warming up connection pool");
 
-        let count = count.min(self.inner.status().max_size as usize);
+        let count = count.min(self.inner.status().max_size);
         let mut connections = Vec::with_capacity(count);
 
         // Acquire connections to force establishment
@@ -171,7 +171,7 @@ impl PgPool {
             "Warming up connection pool with prepared statements"
         );
 
-        let count = count.min(self.inner.status().max_size as usize);
+        let count = count.min(self.inner.status().max_size);
         let mut connections = Vec::with_capacity(count);
 
         for i in 0..count {
